Refactor routes by extracting SSE and link preview queue services
